# Optimize RPC pubsub for multiple clients with the same subscription