--- a/core/src/validator.rs
+++ b/core/src/validator.rs
@@ -471,12 +471,12 @@ impl Validator {
         let optimistically_confirmed_bank =
             OptimisticallyConfirmedBank::locked_from_bank_forks_root(&bank_forks);
 
-        let rpc_subscriptions = Arc::new(RpcSubscriptions::new_with_broadcast_channel_capacity(
+        let rpc_subscriptions = Arc::new(RpcSubscriptions::new_with_config(
             &exit,
             bank_forks.clone(),
             block_commitment_cache.clone(),
             optimistically_confirmed_bank.clone(),
-            config.pubsub_config.queue_capacity,
+            &config.pubsub_config,
         ));
 
         let max_slots = Arc::new(MaxSlots::default());