--- a/replica-node/src/replica_node.rs
+++ b/replica-node/src/replica_node.rs
@@ -208,6 +208,17 @@ fn start_client_rpc_services(
         ));
     }
 
+    let (trigger, pubsub_service) = PubSubService::new(
+        replica_config.pubsub_config.clone(),
+        &subscriptions,
+        replica_config.rpc_pubsub_addr,
+    );
+    replica_config
+        .replica_exit
+        .write()
+        .unwrap()
+        .register_exit(Box::new(move || trigger.cancel()));
+
     let (_bank_notification_sender, bank_notification_receiver) = unbounded();
     (
         Some(JsonRpcService::new(
@@ -231,12 +242,7 @@ fn start_client_rpc_services(
             leader_schedule_cache.clone(),
             max_complete_transaction_status_slot,
         )),
-        Some(PubSubService::new(
-            replica_config.pubsub_config.clone(),
-            &subscriptions,
-            replica_config.rpc_pubsub_addr,
-            &exit,
-        )),
+        Some(pubsub_service),
         Some(OptimisticallyConfirmedBankTracker::new(
             bank_notification_receiver,
             &exit,