--- a/replica-node/src/replica_node.rs
+++ b/replica-node/src/replica_node.rs
@@ -248,7 +248,7 @@ fn start_client_rpc_services(
             &exit,
             bank_forks.clone(),
             optimistically_confirmed_bank.clone(),
-            subscriptions.clone(),
+            subscriptions,
             None,
         )),
     )