#### Problem

Current implementation generates notifications separately for each client, even if they use identical subscriptions. Each client also has a separate outgoing data buffer. This results in suboptimal performance and high memory usage. It also significantly limits the number of connections that can be served at the same time.

#### Summary of Changes

Subscriptions are delivered to clients using a broadcast queue. Each notification for a subscription is only generated and serialized once for all clients.

Implementation notes:

- soketto is used as the WebSocket implementation, but `jsonrpc_core` is still used for parsing requests and generating responses.
- `SubscriptionsTracker` serves the same purpose as old `solana_rpc::rpc_subscriptions::Subscriptions` type, but is modified to make deduplication of subscriptions possible. It provides indexes for efficient selection of relevant subscriptions (by params, by signature, etc.) In addition, it ensures automatic unsubscription when the last subscription token is dropped.
- `notify_accounts_logs_programs_signatures` now does a single iteration over the hashmap. Old implementation collected clones of all the keys into a vector and then did a separate lookup for each key.
- Some simplification and deduplication was made by unifying all subscription parameters into a single type (`SubscriptionParams`) and implementing single subscribe/unsubscribe functions instead of separate functions for each subscription type. That will make adding new subscription types much easier.
- Default values of optional parameters are now exclusively handled in `RpcSolPubSubImpl`. The rest of the code deals with concrete values of all the parameters.

#### Compatibility notes

- Subscription ID is now global for all clients. New subscription ID is generated using a global counter instead of a per-client counter. If a client makes the same subscription twice within a connection, the same subscription ID will be returned.
- Key order in notification values now follows field declaration order. Previous implementation produced sorted keys because of an intermediate conversion to `serde_json::Value`.