@00nktk 2021-08-03
> Thanks! Generally, I'm liking the look of this approach, especially the cleanup in rpc_subscriptions. Have you done any benchmarking or other testing that could give us a concrete idea of the perf improvements?
> If you can rebase on master, we can also give it a spin on a public cluster.

Thanks for the review!
We did test the correctness in terms of consistency with the documented API, as well as somewhat subjective perfomance testing.
Regarding bugs/non-desirable behaviour that got fixed:
- Disconnection problem as described in #18670.
- PubSub lagging behind RPC within the same node.

We're currently running a number of tests, as well as mainnet nodes with this patch. Folks from our testing team will provide a better overview of the progress. @Fedor-sa please leave a comment.