@Fedor-sa 2021-08-03
As for the testing part, we executed several tests comprising:
- Multiple connections (bring up 1000+ unique socket connections, subscribe in each connection separately for the same account, once message received compare messages to each other ensuring all messages are identical)
- WS/RPC consistency (subscribe to WS, receive notification, immediately call RPC method for the same account, compare received date taking into account slot). This test allows us to verify, that WS notifications are not very late. And even if slot in WS notification is behind slot, received upon RPC call, ensure that WS gets up to speed and reduces this gap
- Generic regression testing - other subscription methods, various parameters etc. In fact this issue was found during regression testing https://github.com/solana-labs/solana/issues/18929
Tests are currently in progress
- 50k+ simultaneous subscription (goal is to check validator load)
- 'heavyweight' subscription concurrent with dozen others. Ensure that 'heavyweight' subscription (e.g. programAccount with numerous accounts) does not impact other subscription performance 
You welcome if you have any questions
Fedor