@CriesofCarrots 2021-08-03
> _Tests are currently in progress_
> 
>     * 50k+ simultaneous subscription (goal is to check validator load)
> 
>     * 'heavyweight' subscription concurrent with dozen others. Ensure that 'heavyweight' subscription (e.g. programAccount with numerous accounts) does not impact other subscription performance

This is great! I'm looking forward to seeing the data.
How did the WS/RPC consistency test compare to the current implementation?