@CriesofCarrots 2021-08-05
> > If you can rebase on master, we can also give it a spin on a public cluster.
> 
> Hey @CriesofCarrots, I rebased the branch, we think it would be great to give it a try. Anything else you need to make this happen?

Perfect, thank you! I've got a node set up, so I'll give it a test over the next couple days.