@Fedor-sa 2021-08-06
From testing point of view
-  1000+ connections :white_check_mark:
- 40000+ simultaneous subscriptions `accountSubscribe` in 10 connections :white_check_mark: (also combined with additional "heavyweight" program subscription)
- Data consistency
    1.  WS data vs RPC data on patched validator - :white_check_mark: - this means that there are only small (1-5 Slots) lags observed between WS notification and RPC responses for single pubkey and WS reduces this gap while account data is not being changed rapidly. This test was done with `encoding: base64+zstd`  since data size for `base64` is much larger 
    2. Data  comparison between Mainnet and patched validator - *in progress* 
We found few cases of improvement  (see below), but still trying to check that all data is fully identical for the same slot and `commitment: finalized`


_Mainnet_ 
```
WS SLOT:  90230152 | First element in section "data": 6ad5ebc0e0ed70d27eb5c5c352b5937a |  10:26:15.443227 | api.mainnet-beta.solana.com
RPC SLOT:  90230152 | First element in section "data": 6ad5ebc0e0ed70d27eb5c5c352b5937a |  10:26:16.040630 | api.mainnet-beta.solana.com

WS SLOT:  90230155 | First element in section "data": c37e186ee30bba237c866e1aa7c666f1 |  10:26:17.069676 | api.mainnet-beta.solana.com
RPC SLOT:  90230156 | First element in section "data": 6ad5ebc0e0ed70d27eb5c5c352b5937a |  10:26:18.070193 | api.mainnet-beta.solana.com
```

_Updated validator_
```
WS SLOT:  90230139 | First element in section "data": 6ad5ebc0e0ed70d27eb5c5c352b5937a |  10:26:01.795825 | pub-validator-01.zubr.one:8900
RPC SLOT:  90230139 | First element in section "data": 6ad5ebc0e0ed70d27eb5c5c352b5937a |  10:26:02.004863 | pub-validator-01.zubr.one:8899

WS SLOT:  90230152 | First element in section "data": c37e186ee30bba237c866e1aa7c666f1 |  10:26:15.429553 | pub-validator-01.zubr.one:8900
RPC SLOT:  90230152 | First element in section "data": c37e186ee30bba237c866e1aa7c666f1 |  10:26:15.638836 | pub-validator-01.zubr.one:8899

WS SLOT:  90230156 | First element in section "data": 4d35d0f10d32536237e41886f0899a16 |  10:26:17.732359 | pub-validator-01.zubr.one:8900
RPC SLOT:  90230156 | First element in section "data": 4d35d0f10d32536237e41886f0899a16 |  10:26:17.937133 | pub-validator-01.zubr.one:8899
```

Comments/questions are welcome