@Riateche 2021-08-24
We've done some further improvements to the implementation:
- The pubsub server thread is now more decoupled from the notification generation thread. Instead of sharing the subscription tracker between the threads and accessing it via a mutex, the subscription tracker is now exclusively managed by the notification thread. The server thread pushes deduplicated subscriptions and unsubscriptions to the notification queue. This eliminates contention between the threads.
- There is now a limit on the total memory usage for items in the broadcast queue. 

We've added more metrics and are doing more extensive performance testing. We'll share the results once they're ready.