@giftal 2021-08-24
> Oof, we're deep into testing the previous changeset with real cluster traffic/load; I did not expect more changes of this magnitude.
> I'll hold off on any more until your performance testing is done. Looking forward to seeing the results!
> Please let me know if you expect more sizable changes.



Hey Tyera,

Please go on with your tests, no more changes expected here.