@Riateche 2021-09-01
Not sure what's going wrong on the CI. `cargo test -p solana-rpc` succeeds and takes about a minute total on my system.