@CriesofCarrots 2021-09-02
> Not sure what's going wrong on the CI. `cargo test -p solana-rpc` succeeds and takes about a minute total on my system.

Hmm, okay, I'll try to troubleshoot this tomorrow. What is your rust version?