@Fedor-sa 2021-09-02
Regarding testing: we compared performance between old (1.6.20) and new versions 
_Methodology:_
1. Bring up within 1 minute 1000 connection and send `accountSubscribe` for 1 key in each connection. We used only 4 keys, but we regularly observe notification for them, therefore processing all notifications is a load for ws on validator
2. Keep all connections and subscriptions active for 15 minutes and then disconnect them within 1 minute
_Results_
### Old validator
```
data_received........................: 35 GB   33 MB/s
received_accountNotification_count...: 256090  240.459488/s
ws_msgs_received.....................: 257092  241.400331/s
ws_sessions..........................: 1000    0.938965/s
```
### New validator
```
data_received........................: 153 GB  143 MB/s
received_accountNotification_count...: 801912  752.966887/s
ws_msgs_received.....................: 802919  753.912424/s
ws_sessions..........................: 1000    0.938964/s
```
As we can see from results, validator with new WS sends around 5x times more data and around 3x times more notifications messages in the same test scenario